tmp: removing unused functions: push_to_data_sink and push_to_minio_sink
